refactor: implement suggested changes
