+++ tests/integration/data_loading/es/conftest.py
@@ -1,58 +1,13 @@
 import json
 import os
-from time import sleep
 from warnings import filterwarnings, catch_warnings
 
 import pytest
 from urllib3.exceptions import InsecureRequestWarning, SecurityWarning
 from now.data_loading.es import ElasticsearchConnector, ElasticsearchExtractor
 from tests.integration.data_loading.es.example_dataset import ExampleDataset
-import requests
 
-MAX_RETRIES = 20
-
-
-@pytest.fixture(scope="session")
-def es_connection_params():
-    user_name = 'elastic'
-    password = 'elastic'
-    connection_str = f'https://{user_name}:{password}@localhost:9200'
-    connection_args = {'verify_certs': False}
-    return connection_str, connection_args
-
-
-@pytest.mark.docker
-@pytest.fixture(scope='session', autouse=True)
-def setup_service_running(es_connection_params) -> None:
-    os.system('docker-compose -f tests/resources/text+image/docker-compose.yml up -d')
-    es_connection_str, _ = es_connection_params
-    with catch_warnings():
-        filterwarnings('ignore', category=InsecureRequestWarning)
-        connection_str = es_connection_str
-        retries = 0
-        while True:
-            try:
-                retries += 1
-                if retries > MAX_RETRIES:
-                    raise RuntimeError(
-                        f'Maximal number of retries ({MAX_RETRIES}) reached for '
-                        f'connecting to {connection_str}'
-                    )
-                r = requests.get(connection_str, verify=False)
-                if r.status_code in [200, 401]:
-                    break
-                else:
-                    print(r.status_code)
-                    sleep(5)
-            except Exception:
-                if retries > MAX_RETRIES:
-                    raise RuntimeError(
-                        f'Maximal number of retries ({MAX_RETRIES}) reached for '
-                        f'connecting to {connection_str}'
-                    )
-                sleep(3)
-    yield
-    os.system('docker-compose -f tests/resources/text+image/docker-compose.yml down')
+from tests.integration.data_loading.es.utils import delete_es_index
 
 
 @pytest.fixture
@@ -114,7 +69,7 @@ def setup_online_shop_db(setup_elastic_db, es_connection_params, online_shop_res
     yield es_connector, index_name
 
     # delete index
-    es_connector.delete_index(name=index_name)
+    delete_es_index(connector=es_connector, name=index_name)
 
 
 @pytest.fixture
